Refactor:Toast 컴포넌트 Footer에 종속시키지 않고 ToastProvider 클라이언트 컴포넌트 생성하여 리팩토링
