+++ app/(pages)/layout.tsx
@@ -1,5 +1,6 @@
 import RecoilRootLayout from "@common/components/layout/RecoilRootLayout/RecoilRootLayout";
 import "../styles/global.css";
+import ToastProvider from "@common/components/layout/ToastProvider/ToastProvider";
 
 export const metadata = {
   title: "로컬무드",
@@ -15,7 +16,10 @@ export default async function RootLayout({
     <html lang="en" className="width-[100%] height-[100%]">
       <body className="w-[100%] h-[100%]">
         <main className="w-[100%] h-[100%] fixed overflow-hidden">
-          <RecoilRootLayout>{children}</RecoilRootLayout>
+          <RecoilRootLayout>
+            {children}
+            <ToastProvider />
+          </RecoilRootLayout>
         </main>
       </body>
     </html>
